feat(add new bots form): add new bots form
