+++ src/components/AddNewBotSection/AddNewBotForm/AddNewBotForm.module.scss
@@ -1,0 +1,117 @@
+@import './../../../vendor/_varibles.scss';
+
+.form {
+    max-width: 1320px;
+    width: 100%;
+    display: grid;
+    grid-template-columns: minmax(150px, 240px) repeat(2, minmax(325px, 510px));
+    column-gap: 30px;
+    row-gap: 20px;
+
+  &__label {
+    display: flex;
+    flex-direction: column;
+    row-gap: 5px;
+    padding: 0 0 23px;
+
+    &_type_name {
+      grid-column-start: 1;
+      grid-column-end: 3;
+    }
+
+    &_type_category {
+      grid-column-start: 1;
+      grid-column-end: 3;
+    }
+
+    &_type_exaples {
+      grid-column-start: 1;
+      grid-column-end: 5;
+    }
+  }
+
+  &__inputTitle {
+    padding: 0 0 0 10px;
+    @include text(
+      $font-small-14,
+      $weight-300,
+      $line-height-small-14,
+      $color-blue-dark-title
+    );
+  }
+
+  &__inputGroup {
+    padding: 0;
+    border: none;
+    display: grid;
+    grid-template-rows: repeat(3, 1fr);
+    grid-template-columns: repeat(2, 1fr);
+    column-gap: 32px;
+    row-gap: 20px;
+  }
+
+  &__input {
+    width: 100%;
+    height: 50px;
+    border: $border-main;
+    border-radius: 10px;
+    padding: 0 0 0 20px;
+    background-color: $background-input;
+    @include text(
+      $font-small-14,
+      $weight-300,
+      $line-height-small-14,
+      $color-blue-input-text
+    );
+
+    &_type_desc {
+      height: 100%;
+    }
+
+    &_type_exaples {
+      height: 200px;
+    }
+
+    &::placeholder {
+      @include text(
+        $font-small-14,
+        $weight-300,
+        $line-height-small-14,
+        $input-placeholder
+      );
+    }
+
+    &:focus {
+      outline: none;
+    }
+  }
+
+  &__deleteBtn {
+    border: none;
+    background-color: $background-white;
+    text-align: start;
+    @include text(
+    $font-medium-24,
+    $weight-400,
+    $line-height-medium-24,
+    $color-red-delete
+    );
+    cursor: pointer;
+    transition: opacity .3s;
+
+    &::before {
+        content: '';
+        display: inline-block;
+        width: 24px;
+        height: 24px;
+        background-image: url('../../../images/trash-icon.svg');
+        background-repeat: no-repeat;
+        background-size: contain;
+        padding-right: 10px;
+    }
+
+    &:hover {
+        opacity: .7;
+    }
+  }
+}
