+++ src/components/App/App.jsx
@@ -21,6 +21,7 @@ import { fetchInitialBots, fetchSearchBots } from '../../utils/api/getBots';
 import * as authorizeApi from '../../utils/api/authorizeApi';
 import * as userApi from '../../utils/api/userApi';
 import RegisterSeller from '../RegisterSeller/RegisterSeller';
+import AddNewBotsPage from '../AddNewBotsPage/AddNewBotsPage';
 
 const App = () => {
   const navigate = useNavigate();
@@ -289,6 +290,10 @@ const App = () => {
           <Route
             path='/signup-seller'
             element={<RegisterSeller comeBack={handleGoBack} />}
+          />
+          <Route
+            path='/add-new-bots'
+            element={<AddNewBotsPage comeBack={handleGoBack} />}
           />
         </Routes>
         <Footer />
