+++ src/components/Cart/Cart.jsx
@@ -20,7 +20,7 @@ function Cart({
   decreaseProductCount,
   comeBack,
 }) {
-  const [totalSum, setTotalSum] = useState(0.0); // состояние для общей суммы заказа
+  const [totalSum, setTotalSum] = useState(0); // состояние для общей суммы заказа
 
   // функция нахождения общего е=количнства товаров в корзине
   const count = () => {
@@ -45,21 +45,24 @@ function Cart({
   // функция расчета общей суммы заказа
   const findTotalSum = useCallback(() => {
     return cartProducts.reduce((previousValue, product) => {
-      return (
-        previousValue + parseFloat(product.price).toFixed(2) * product.count
-      );
-    }, 0.0);
+      return previousValue + product.price * product.count;
+    }, 0);
   }, [cartProducts]);
 
   useEffect(() => {
     const sum = findTotalSum();
-    setTotalSum(sum.toFixed(2));
+    setTotalSum(sum);
   }, [cartProducts, findTotalSum]);
+
+  // прокрутка скролла наверх
+  useEffect(() => {
+    window.scrollTo(0, 0);
+  }, []);
 
   return (
     <section className={styles.cart}>
       <div className={styles.products}>
-        <BackButton currentPageName='Корзина' comeBack={comeBack} />
+        <BackButton comeBack={comeBack} />
         <ul className={styles.products__list}>
           {cartProducts.map((product) => (
             <CartProduct
